+++ contribs/common/src/main/java/org/matsim/contrib/common/conventions/vsp/SubpopulationDefaultNames.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.vsp.scenario;
+package org.matsim.contrib.common.conventions.vsp;
 
 public class SubpopulationDefaultNames{
 	/**
